[core] split TestEnv TypedEncoded types into their own module
